requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
python-dotenv==1.0.0
anthropic==0.7.7
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
neo4j==5.14.0
networkx==3.2.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
plotly==5.18.0
pyvis==0.3.2
fpdf2==2.7.6
httpx==0.25.2
python-multipart==0.0.6
loguru==0.7.2

